explicitly apply GAS_LIMIT to shutdownSystem()
"If you are using ethers.js, then those values need to be set in each
call/deployment.

"The gasPrice that can be set in the config doesn't have an effect when
you are using ethers. We need to update that section of the docs to make
it clearer that that's the case."

source: https://twitter.com/fvictorio_nan/status/1470503282909093890

I confirmed with fvictorio that the same advice applies to the gas limit
as well.

I confirmed that, before this change, the log from the hardhat benchmark
run showed a call to eth_estimateGas, and after this change it did not.